requirements.txt
djitellopy
opencv-python
openai
python-dotenv
pyaudio